+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,7 +1,7 @@
 import NextAuth from "next-auth/next";
 import GoogleProvider from 'next-auth/providers/google' //OAuth
 import GitHubProvider from "next-auth/providers/github";
-
+import CredentialsProvider from "next-auth/providers/credentials";
 // export default NextAuth({
 //     providers: [
 //         GoogleProvider({
@@ -29,14 +29,25 @@ const handler = NextAuth({
         GitHubProvider({
             clientId: process.env.GITHUB_ID!,
             clientSecret: process.env.GITHUB_SECRET!,
-          })
-    ],
-    //this will overwrite my callbackUrl in login page.tsx
-    // callbacks: {
-    //     async redirect({ url, baseUrl}){
-    //         return baseUrl;
-    //     }
-    // }
-})
+        }),
+        // CredentialsProvider({
+        //   type: 'credentials',
+        //   credentials: {},
+        //   authorize(credentials, req) {
+        //     const { email, password } = credentials as { 
+        //       email: string;
+        //       password: string; 
+        //     }
 
-export {handler as GET, handler as POST};
+        //   }
+        // })      //         STILL WORKING ON THIS TO ADD CLIENT BASE ON EMAIL AND PASSWORD WITHOUT OATH  DONT DELETE JUST YET `
+    ],
+  })
+  
+  export {handler as GET, handler as POST};
+  //this will overwrite my callbackUrl in login page.tsx
+  // callbacks: {
+  //     async redirect({ url, baseUrl}){
+  //         return baseUrl;
+  //     }
+  // }
