+++ src/app/components/login/page.tsx
@@ -1,23 +1,19 @@
-'use client'
-import {signIn, signOut, useSession} from 'next-auth/react';
+import { NextPage } from 'next';
 
+interface Props{}
 
-export default function Login(){
-    const { data: session } = useSession();
-    if(session){
-    return(
-        <div>
-            <button onClick={() => signOut()} type="button" className='btn btn-primary'>
-                SignOut
-            </button>
-        </div>
-        )
-    }
-        return(
-            <div>
-                <button onClick={() => signIn(undefined, { callbackUrl: 'http://localhost:3000/components/home'}) } type="button" className='btn btn-primary'>
-                    SignIn
-                </button>
-            </div>
-            )
-};
+const Login: NextPage = (props): JSX.Element => {
+  return (
+    <div className='sign-in-form'> 
+      <form>
+        <h1>LOGIN</h1>
+        <input type="email" placeholder='john@email.com' />
+        <input type="password" placeholder='**********' />
+        <input type="submit" value="Login" />
+
+      </form>
+    </div>
+  )
+}
+
+export default Login
