+++ src/app/components/signUp/page.tsx
@@ -1,7 +1,18 @@
-import React from 'react';
+import { NextPage } from 'next';
 
-export default function Something(){
-    return(
-        <div>Sign Up</div>
-    )
-};
+interface Props{}
+
+const Signup: NextPage = (props): JSX.Element => {
+  return (
+    <div className='sign-in-form'> 
+      <form>
+        <h1>SIGNUP</h1>
+        <input type="email" placeholder='john@email.com' />
+        <input type="password" placeholder='**********' />
+        <input type="submit" value="Signup" />
+      </form>
+    </div>
+  )
+}
+
+export default Signup
